Fix Dropdown for Rich Text Editor in the inspector
When editing a rich text editor content in the inspector, as soon as content is changed the element is promoted to a higher z-index for unpublished changes. This results in the standard dropdown contents being hidden behind the editor. This updates the dropdown z-index to use the --zIndex-SecondaryInspectorElevated-DropDownContents for the dropdown contents to fix the issue.